requirements.txt
torch
torchvision
transformers
pillow
matplotlib
sentence-transformers
numpy
scikit-learn